Start of pipe support for zero tier.
Rethinking a little bit how we structure the pipe negotiation code,
we are going to create the pipes right away for connect and accept
(once we have gotten past bringing the network up at least), and
then run the 3-way negotiation from the pipe.

While here, we noticed that we were allocating ports from the endpoint
hash, rather than a global port hash.  That would be a problem once
we start opening up pipes.  So now we allocate from a new ports hash.

+++ src/nng.h
@@ -638,7 +638,7 @@ struct nng_sockaddr_in {
 struct nng_sockaddr_zt {
 	uint64_t sa_nwid;
 	uint64_t sa_nodeid;
-	uint16_t sa_port;
+	uint32_t sa_port;
 };
 
 typedef struct nng_sockaddr_in nng_sockaddr_in;
